had to change map uploads to be stored at the storage rather than firestore. upload now works, need to fetch uploads correctly

+++ app/src/main/java/com/technion/columbus/main/MainActivity.kt
@@ -132,7 +132,7 @@ class MainActivity : AppCompatActivity() {
                 return@positiveButton
             }
 
-            val scanRadius = seekBarProgressToScanRadius(scanRadiusSeekBar.progress)
+            val scanRadius = seekBarProgressToScanRadius(scanRadiusSeekBar.progress).toFloat()
 
             val startGameIntent = Intent(this@MainActivity, GameMapActivity::class.java)
             startGameIntent.putExtra(SCAN_NAME, name.toString())
